+++ CHANGELOG.md
@@ -6,6 +6,7 @@ We follow [Semantic Versions](https://semver.org/).
 
 - Make `Dependency.__call__` async because FastAPI always runs Picodi deps in threadpool even if they are async
 - Rename `Dependency` to `Depends` [internal change]
+- Run mypy in tests directory
 
 ## Version 0.20.0